The order of the page sections has been changed, now it is: 1. Hero 2. Services 3. Portfolio 4. Team 5. Reviews 6. Contact
The services section has been created with brochures describing the services the company provides.

The portfolio section is being worked on, where for now the title and subtitle have been changed, and a video component has been added.

The theme and colors for the entire page have been improved.

The team section has been improved by changing some aspects of the slider.

Some colors of the dark theme have been changed, and the secondary color has been changed to a darker blue.

The footer has been improved and changed.